backend/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.28
Werkzeug==3.0.1
selenium==4.18.1
schedule==1.2.1
webdriver-manager==4.0.1
